# docs/swagger.yaml
basePath: /api
definitions:
...
  ds.BiomarkerItemResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      max_value:
        type: number
      measure_unit:
        type: string
      min_value:
        type: number
      name:
        type: string
      patient_value:
        type: number
      significance:
        type: number
    type: object
  ds.INIResearchStatus:
    enum:
    - черновик
    - удалён
    - сформирован
    - завершён
    - отклонён
    type: string
    x-enum-varnames:
    - StatusDraft
    - StatusDeleted
    - StatusFormed
    - StatusCompleted
    - StatusRejected
...
  ds.ResearchWithBiomarkersResponse:
    properties:
      biomarkers:
        description: Биомаркеры с нужными полями для фронтенда
        items:
          $ref: '#/definitions/ds.BiomarkerItemResponse'
        type: array
      completed_at:
        type: string
      created_at:
        type: string
      creator_login:
        type: string
      formed_at:
        type: string
      id:
        type: integer
      ini_result:
        type: number
      moderator_login:
        type: string
      patient_birth:
        type: string
      patient_gender:
        type: string
      patient_name:
        type: string
      status:
        $ref: '#/definitions/ds.INIResearchStatus'
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@nutriscan.ru
    name: API Support
    url: http://localhost:8081
  description: API для расчета индекса нутритивной недостаточности
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: NutriScan API
  version: "1.0"
paths:
...
  /researches:
    get:
      description: Возвращает список исследований с возможностью фильтрации по статусу
        и дате
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: 'Начальная дата (формат: 2006-01-02)'
        in: query
        name: start_date
        type: string
      - description: 'Конечная дата (формат: 2006-01-02)'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список исследований
      tags:
      - Исследования
  /researches/{id}:
    delete:
      description: Удаляет исследование (только создатель для черновиков)
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Исследование удалено
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить исследование
      tags:
      - Исследования
    get:
      description: Возвращает детальную информацию об исследовании с биомаркерами
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/ds.ResearchWithBiomarkersResponse'
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить исследование по ID
      tags:
      - Исследования
    put:
      consumes:
      - application/json
      description: Обновляет информацию о пациенте в исследовании (только для врачей)
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      - description: Информация о пациенте
        in: body
        name: patient_info
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Исследование обновлено
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить информацию об исследовании
      tags:
      - Исследования
  /researches/{id}/complete:
    put:
      consumes:
      - application/json
      description: Завершает или отклоняет сформированное исследование (только для
        модераторов)
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      - description: Действие
        in: body
        name: action
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Исследование завершено/отклонено
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить или отклонить исследование
      tags:
      - Исследования
  /researches/{id}/form:
    put:
      description: Переводит исследование из статуса черновика в сформированный (только
        создатель)
      parameters:
      - description: ID исследования
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Исследование сформировано
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные или пустое исследование
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Исследование не найдено
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать исследование
      tags:
      - Исследования
  /researches/cart:
    get:
      description: Возвращает ID черновика и количество биомаркеров в корзине
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить информацию о корзине исследований
      tags:
      - Исследования
...
securityDefinitions:
  ApiKeyAuth:
    description: Введите "Bearer [ваш_JWT_токен]" для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...